fix semantic_errors CSS
My goal here is to remove the extra vertical space occurring after the
ul.error-list element. The source of the space, from Bootstrap, is:

ul, ol {
  padding: 0;
  margin: 0 0 10px 25px; }

The simplest change is to modify `margin-bottom`. The code that I'm
replacing, `margin: 0` pressed the list bullets against the left of the
alert box with no margin.