Adds "form-control" class where required.
This is work to support Bootstrap 3.  In this version "form-control"
is added.

I needed to do some interpretation about how datetime_select and
friends should work.  It sort of works now, but only on FireFox:
WebKit browsers seem to ignore the "height" attribute added in the
CSS.  I might revisit this at some point -- it's pretty annoying.